Rename `lib/$app/env` to `lib/$app/environment` (SvelteKit change)
