+++ src/lib/$app/environment/index.ts
@@ -1,0 +1,3 @@
+// Mimic SvelteKit's `$app/environment` foor non-SSR outputs in RollUp only and Vite only setups.
+
+export const browser = true
